+++ src/TechLanches/TechLanches.Domain/Services/IProdutoService.cs
@@ -7,8 +7,8 @@ namespace TechLanches.Domain.Services
         Task<List<Produto>> BuscarTodos();
         Task<Produto> BuscarPorId(int produtoId);
         Task<List<Produto>> BuscarPorCategoria(int categoriaId);
-        Task<Produto> Cadastrar(string nome, string descricao, double preco, int categoriaId);
-        Task Atualizar(int produtoId, string nome, string descricao, double preco, int categoriaId);
+        Task<Produto> Cadastrar(string nome, string descricao, decimal preco, int categoriaId);
+        Task Atualizar(int produtoId, string nome, string descricao, decimal preco, int categoriaId);
         Task Deletar(int produtoId);
     }
 }
